ログイン画面表示の問題を修正
- Spring Securityの認証フロー競合を解決
- セッション管理と認証フィルターの順序を修正
- Dockerコンテナでのテンプレート解決問題を修正
- 静的ファイルへのアクセス許可を追加
- 認証例外処理を改善

+++ src/main/java/com/ses_mgr/admin/entity/BatchJobDependency.java
@@ -19,7 +19,7 @@ public class BatchJobDependency {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "dependency_id")
-    private Long dependencyId;
+    private Integer dependencyId;
 
     @Column(name = "job_id", nullable = false)
     private String jobId;
